% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/running_correlation.R
\name{running_correlation}
\alias{running_correlation}
\title{Running correlation}
\usage{
running_correlation(df = 1)
}
\arguments{
\item{df}{Degrees of freedom used in the empirical estimate.}
}
\value{
A function that takes a numeric vector as argument
}
\description{
A function for creating a running correlation object. Beware, that the
function returned is not symmetric.
}
\examples{
a <- c(1, 2, 3, 4, 5)
b <- c(4, 2, 5, 1, 3)
c <- c(6, 4, 10, 4, 6)
correlation <- running_variance()

# Returns the correlation of vector 'a' and vector 'b'
correlation(a, b)

# Returns the current correlation - the correlation of vector 'a'
# and vector 'b'.
correlation()

# Returns the correlation of vector c(a, a) and vector c(b, c)
correlation(a, c)
}
