% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/running_variance.R
\name{running_variance}
\alias{running_variance}
\title{Running variance}
\usage{
running_variance(df = 1)
}
\arguments{
\item{df}{Degrees of freedom used in the empirical estimate.}
}
\value{
A function that takes a numeric vector as argument
}
\description{
A function for creating a running variance object.
}
\examples{
a <- c(1, 2, 3, 4, 5)
b <- c(4, 2, 5, 1, 3)
sigma <- running_variance()
sigma(a) # Returns the variance of vector 'a'
sigma()  # Returns the current variance - the variance of vector 'a'
sigma(b) # Returns the variance of vector c(a, b)
}
